Pin wheel in securedrop-app-code requirements
Remove docker-requirements files as pip, setuptools, and wheel are now
all pinned in the securedrop-app-code requirements, which are used in
all the same places.

Use a virtualenv for prepackaging tasks like compiling translations in
the container used to build the securedrop-app-code Debian package.

Fix a missing variable (due to copypasta) that could break retrieval
of the finished securedrop-app-code deb from the build container.